refactor(workflows): dedup helpers, wire retention, narrow journal kind, share run wire shape
